Rename files to reflect command name 'ManageRemote'
