+++ src/style/ManageRemoteDialog.ts
